Update Ogre2RayQuery.cc
made changes to the ClosestPointByIntersection method. Earlier the method would filter the visual objects first based on the Oriented bounding box. This was an issue as if we had the robot inside another bounding box the ray would just pass through. This change removes the fiter, now the ray will chekc collision with all the objects even if the Oriented Bounding boxes are intersecting

Signed-off-by: jacobrejin <jacobrejin@gmail.com>